% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funcs.r
\name{ss_category}
\alias{ss_category}
\title{Get Saffir Simpson category given max wind speed.}
\usage{
ss_category(vx, units = "kt")
}
\arguments{
\item{vx}{The max wind speed, in knots}

\item{units}{Character vector, either "kt" (default),
"m/s", or "mph"}
}
\value{
cat Category/ies. Values 1-5 denote category 1-5
hurricanes; 0 denotes tropical storm; -1 denotes tropical
depression.
}
\description{
\code{ss_category} returns to Saffir-Simpson Category/ies of
a tropical cyclone given the (vector of) maximum wind speed(s).
}

