% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funcs.r
\name{TornSnyder2012_pmin_uncert}
\alias{TornSnyder2012_pmin_uncert}
\title{Get modern-day central pressure uncertainty estimates from Torn & Snyder 2012.}
\usage{
TornSnyder2012_pmin_uncert(pmin = NA, SScat = NA, vmax = NA)
}
\arguments{
\item{pmin}{(Vector of) central pressure(s) of the storm(s), in hPa.}

\item{SScat}{Alternately, can input the Saffir-Simpson category
(-1 for tropical depression; 0 for tropical storm, 1-5 for Cat 1-5 storms.)}

\item{vmax}{Alternatively, (vector of) max wind speed(s) of the TC(s) in knots.}
}
\value{
sig Uncertainty estimate in hPa; of same size as input vector.
}
\description{
\code{TornSnyder2012_pos_uncert} returns the estimates of intensity-dependent
uncertainty in maximum TC windspeed from the study:
Torn, R.D., and C.M. Snyder, 2012: Uncertainty of tropical cyclone best-track
information. Weather and Forecasting, 27, 715-729, DOI: 10.1175/WAF-D-11-00085.1.
}

