% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funcs.r
\name{TornSnyder2012_vx_uncert}
\alias{TornSnyder2012_vx_uncert}
\title{Get modern-day observational max wind speed uncertainty estimates from Torn & Snyder 2012.}
\usage{
TornSnyder2012_vx_uncert(vmax = NA, units = "kts", SScat = NA)
}
\arguments{
\item{vmax}{(Vector or scalar of) max wind speed(s) (default: NA).}

\item{units}{Character vector giving the units of the input windspeed
(default: "kts". Can also be "m/s" or "mph").}

\item{SScat}{Alternately, can input the Saffir-Simpson category
(-1 for tropical depression; 0 for tropical storm, 1-5 for Cat 1-5 storms.)}
}
\value{
sig Uncertainty estimate in knots; of same size as input vector.
}
\description{
\code{TornSnyder2012_vx_uncert} returns the estimates of intensity-dependent
uncertainty in maximum TC windspeed from the study:
Torn, R.D., and C.M. Snyder, 2012: Uncertainty of tropical cyclone best-track
information. Weather and Forecasting, 27, 715-729, DOI: 10.1175/WAF-D-11-00085.1.
}

